Port gpt2 (#23)
* Initialized modules copied from HuggingFace's GPT2 code.

* Added attention module from HuggingFace.

* Made placeholder file / object for the custom positional embeddings.

* Reorganized to new directory structure.

* Origin Merge

* Fixed ids for tokens to match the data.

* Changed mechanism for getting possible wait tokens.

* Added gpt2 model.

* Fixed not registered error.

* Added 3 missing tokens.

* Fixed the cross-entropy loss for the baseline model (was negative)

* Removed redundant file.

* chunk instance

* Fixed sequence len issue. Needs to be done for the validation loop too.

* Added changes to the validation loop as well.

* Added gpt2 model.

* Fixed typos.

* Made the frequency weights assigned through a  buffer.

* Fixed bugs in training loop as well as added a CUDA device selection arg

* Fixed typo

* Removed build folder

* Added default chunk_size to training

* Updated tests

* Minor formatting fixes

* Fixed to data to handle any possible note under new quantization settings

* Added gradient accumulation, gradient checkpointing, and fp16 support

* Registered optimizer from pytorch_pretrained_bert

* Added requirements to appease CI

* Proper apex install in CI

+++ squawkbox/commands/train.py
@@ -7,6 +7,7 @@ from pathlib import Path
 import shutil
 import sys
 
+from apex import amp
 import numpy as np
 import torch
 from torch.utils.data import DataLoader
@@ -23,23 +24,32 @@ logger = logging.getLogger(__name__)
 
 def _train(args):
     """Training function"""
-    # Front matter
+    # Front matter, fail early if possible
+    if args.data_parallel and not args.cuda:
+        logger.error('Cannot use --data-parallel without --cuda. Exiting.')
+        sys.exit(1)
+
+    if not args.config.exists():
+        logger.error('Config does not exist. Exiting.')
+        sys.exit(1)
+
+    with open(args.config, 'r') as config_file:
+        config = yaml.load(config_file, Loader=yaml.SafeLoader)
+
     if args.output_dir.exists() and not args.resume:
-        logger.error('Directory "%s" already exists. Exiting.')
+        logger.error('Directory "%s" already exists. Exiting.' % str(args.output_dir))
         sys.exit(1)
     else:
         logger.info('Creating directory "%s"', args.output_dir)
         args.output_dir.mkdir()
         shutil.copy(args.config, args.output_dir / 'config.yaml')
 
-    with open(args.config, 'r') as config_file:
-        config = yaml.load(config_file, Loader=yaml.SafeLoader)
-
     torch.manual_seed(config.get('seed', 5150))
     np.random.seed(config.get('seed', 1336) + 1)
 
     # Initialize model components from config
     model = Model.from_config(config['model'])
+
     optimizer = Optimizer.from_config(config['optimizer'], params=model.parameters())
     if 'lr_scheduler' in config:
         lr_scheduler = LRScheduler.from_config(config['lr_scheduler'])
@@ -47,7 +57,15 @@ def _train(args):
         lr_scheduler = None
 
     if args.cuda:
-        model = model.cuda()
+        if args.cuda_device is not None:
+            model = model.cuda(args.cuda_device)
+        elif args.data_parallel and not args.fp16:
+            model = torch.nn.DataParallel(model)
+        else:
+            model = model.cuda()
+
+        if args.fp16:
+            model, optimizer = amp.initialize(model, optimizer, opt_level='O1')
 
     # Restore checkpoint
     checkpoint_path = args.output_dir / 'checkpoint.pt'
@@ -70,6 +88,8 @@ def _train(args):
     validation_dataset = MidiDataset(config['validation_data'])
 
     train_config = config['training']
+    chunk_size = train_config.get('chunk_size', 512)
+    step = 0
     for epoch in range(start_epoch, train_config['epochs']):
         logger.info('Epoch: %i', epoch)
 
@@ -82,14 +102,44 @@ def _train(args):
                                   num_workers=train_config.get('num_workers', 0),
                                   collate_fn=pad_and_combine_instances)
         train_tqdm = tqdm(train_loader, desc='Loss: NA')
+        optimizer.zero_grad()
         for instance in train_tqdm:
             if args.cuda:
                 instance = {key: value.cuda() for key, value in instance.items()}
-            optimizer.zero_grad()
-            output_dict = model(**instance)
-            loss = output_dict['loss']
-            loss.backward()
-            optimizer.step()
+
+            instance_chunks = {key: torch.split(value, chunk_size, dim=1)
+                              for key, value in instance.items()}
+
+            output_dict = {"hidden": None}
+            keep_id_list = [] #[instance_chunk["src"][:, 0] != 0]
+            for chunk_id in range(len(instance_chunks['src'])):
+                instance_chunk = {key: value[chunk_id] for key, value in instance_chunks.items()}
+
+                for keep_id in keep_id_list:
+                    instance_chunk = {key: value[keep_id, :] for key, value in instance_chunk.items()}
+
+                # need to filter empty sequences out
+                new_keep_ids = instance_chunk["src"][:, 0] != 0
+                keep_id_list.append(new_keep_ids)
+
+                instance_chunk = {key: value[new_keep_ids, :] for key, value in instance_chunk.items()}
+
+                if output_dict["hidden"] is not None:
+                    output_dict["hidden"] = [h_vec[:, new_keep_ids, :].detach() for h_vec in output_dict["hidden"]]
+
+                output_dict = model(hidden=output_dict["hidden"], **instance_chunk)
+                loss = output_dict['loss']
+                if args.fp16:
+                    with amp.scale_loss(loss,  optimizer) as scaled_loss:
+                        scaled_loss.backward()
+                else:
+                    loss.backward()
+
+                step += 1
+                if not step % train_config.get('accumulation_steps', 1):
+                    optimizer.step()
+                    optimizer.zero_grad()
+
             train_tqdm.set_description('Loss: %0.4f' % loss.item())
 
         # Validation loop
@@ -105,18 +155,41 @@ def _train(args):
         number_of_instances = 0
         for instance in validation_tqdm:
             if args.cuda:
-                instance = {key: value.cuda() for key, value in instance.items()}
-            output_dict = model(**instance)
-            loss = output_dict['loss']
-            total_validation_loss += loss.item()
-            number_of_instances += train_config['batch_size']
-            validation_tqdm.set_description('Loss: %0.4f' % (total_validation_loss / number_of_instances))
+                instance = {key: value.cuda(args.cudadev) for key, value in instance.items()}
+
+            keep_id_list = [instance["src"][:, 0] != 0]
+            instance_chunks = {key: torch.split(value, chunk_size, dim=1) for key, value in instance.items()}
+            output_dict = {"hidden": None}
+            for chunk_id in range(len(instance_chunks['src'])):
+                instance_chunk = {key: value[chunk_id] for key, value in instance_chunks.items()}
+
+                for keep_id in keep_id_list:
+                    instance_chunk = {key: value[keep_id, :] for key, value in instance_chunk.items()}
+
+                # need to filter empty sequences out
+                new_keep_ids = instance_chunk["src"][:, 0] != 0
+                keep_id_list.append(new_keep_ids)
+
+                instance_chunk = {key: value[new_keep_ids, :] for key, value in instance_chunk.items()}
+
+                if output_dict["hidden"] is not None:
+                    output_dict["hidden"] = [h_vec[:, new_keep_ids, :].detach() for h_vec in output_dict["hidden"]]
+
+                output_dict = model(hidden=output_dict["hidden"], **instance_chunk)
+                loss = output_dict['loss']
+                total_validation_loss += loss.item()
+                number_of_instances += new_keep_ids.squeeze().sum().item() #train_config['batch_size']
+                validation_tqdm.set_description('Loss: %0.4f' % (total_validation_loss / number_of_instances))
         metric = total_validation_loss / number_of_instances
         logger.info('Validation Loss: %0.4f', metric)
 
         # Checkpoint
+        if args.data_parallel:
+            model_state_dict = model.module.state_dict()
+        else:
+            model_state_dict = model.state_dict()
         state_dict = {
-            'model': model.state_dict(),
+            'model': model_state_dict,
             'optimizer': optimizer.state_dict(),
             'epoch': epoch,
             'best_metric': best_metric
@@ -128,6 +201,7 @@ def _train(args):
         if metric < best_metric:
             logger.info('Best model so far.')
             state_dict['best_metric'] = metric
+            best_metric = metric
             torch.save(state_dict, best_checkpoint_path)
 
         # Save current model.
@@ -143,6 +217,11 @@ if __name__ == '__main__':
     parser.add_argument('config', type=Path, help='path to config .yaml file')
     parser.add_argument('output_dir', type=Path, help='output directory to save model to')
     parser.add_argument('--cuda', action='store_true', help='use CUDA')
+    parser.add_argument('--cuda-device', type=int, help='CUDA device num', default=None)
+    parser.add_argument('--data-parallel', action='store_true',
+                        help='If enabled, trains on all GPUs')
+    parser.add_argument('--fp16', action='store_true',
+                        help='Enables half precision training')
     parser.add_argument('-r', '--resume', action='store_true',
                         help='will continue training existing checkpoint')
     args, _ = parser.parse_known_args()
